+++ resources/client.rb
@@ -1,9 +1,16 @@
 actions :create
 
 attribute :address, :kind_of => String, :required => true
-attribute :subscriptions, :kind_of => Array, :default => Array.new
-attribute :keepalive, :kind_of => Hash, :default => Hash.new
-attribute :additional, :kind_of => Hash, :default => Hash.new
+attribute :subscriptions, :kind_of => Array, :default => []
+attribute :keepalive, :kind_of => Hash, :default => {}
+attribute :keepalives, :kind_of => [TrueClass, FalseClass], :default => true
+attribute :safe_mode, :kind_of => [TrueClass, FalseClass], :default => false
+attribute :redact, :kind_of => Array, :default => []
+attribute :socket, :kind_of => Hash, :default => {}
+attribute :registration, :kind_of => Hash, :default => {}
+attribute :deregister, :kind_of => [TrueClass, FalseClass], :default => false
+attribute :deregistration, :kind_of => Hash, :default => {}
+attribute :additional, :kind_of => Hash, :default => {}
 
 def initialize(*args)
   super
@@ -11,7 +18,5 @@ def initialize(*args)
 end
 
 def after_created
-  unless name =~ /^[\w\.-]+$/
-    raise Chef::Exceptions::ValidationFailed, "Sensu client name cannot contain spaces or special characters"
-  end
+  raise Chef::Exceptions::ValidationFailed, "Sensu client name cannot contain spaces or special characters" unless name =~ /^[\w\.-]+$/
 end